refactor: shift enter for message in chat page
