+++ src/modules/chats/components/sendMessage/index.module.scss
@@ -1,0 +1,27 @@
+@import 'src/styles/constants/_primitives.scss';
+
+.box {
+    width: 100%;
+    position: relative;
+    padding: 0.75rem 6rem 0.75rem 0.875rem;
+    border: 1px solid $color-grey-300;
+    border-radius: 8px;
+
+    &:focus-within {
+        box-shadow: 0px 0px 0px 4px $color-primary-100;
+    }
+
+    .input {
+        width: 100%;
+        height: 8.75rem;
+        border: none;
+
+        &::placeholder {
+            color: $color-grey-900;
+            font-size: 16px;
+            font-style: normal;
+            font-weight: 400;
+            line-height: 24px;
+        }
+    }
+}
