+++ src/modules/chats/components/sendMessage/index.tsx
@@ -1,38 +1,33 @@
-import React, { useState } from 'react';
+import { useState } from 'react';
 import { Button } from 'src/modules/general/components/Button';
 import { IconButton } from 'src/modules/general/components/iconButton';
 import { Input } from 'src/modules/general/components/input/input';
 import variables from 'src/styles/constants/_exports.module.scss';
 
-import css from './sendMessage.module.scss';
+import css from './index.module.scss';
+import { SendMessageProps } from './index.types';
 
-interface SendMessageProps {
-  receipientId?: string;
-  onSend?: (message: string) => void;
-  handleCreateChat?: (receipientId: string, text: string) => void;
-}
-export const SendMessage: React.FC<SendMessageProps> = ({ onSend, handleCreateChat, receipientId }) => {
+export const SendMessage: React.FC<SendMessageProps> = ({ onSend, handleCreateChat, recipientId }) => {
   const [newMessage, setNewMessage] = useState('');
 
-  const handleSendMessage = async () => {
-    if (onSend) await onSend(newMessage);
-    else if (handleCreateChat && receipientId) await handleCreateChat(receipientId, newMessage);
+  const handleSendMessage = () => {
+    onSend?.(newMessage);
+    if (recipientId) handleCreateChat?.(recipientId, newMessage);
     setNewMessage('');
   };
 
-  const enterInput = (e: any) => {
-    if (e.keyCode === 13) {
-      handleSendMessage();
-    }
+  const enterInput = (e: React.KeyboardEvent<HTMLTextAreaElement>) => {
+    if (e.key === 'Enter' && !e.shiftKey) handleSendMessage();
   };
+
   return (
     <>
-      <div className={`hidden md:flex ${css.sendBox}`}>
+      <div className={`hidden md:flex ${css['box']}`}>
         <textarea
           value={newMessage}
           onChange={e => setNewMessage(e.target.value)}
-          onKeyDown={e => enterInput(e)}
-          className={css.inputMessage}
+          onKeyDown={enterInput}
+          className={css['input']}
           placeholder="Send a message"
         />
 
@@ -45,13 +40,13 @@ export const SendMessage: React.FC<SendMessageProps> = ({ onSend, handleCreateCh
           Send
         </Button>
       </div>
-      <div className={`flex md:hidden py-6 gap-3`}>
+      <div className="flex md:hidden py-6 gap-3">
         <div className="flex-1">
           <Input
             value={newMessage}
             onChange={e => setNewMessage(e.target.value)}
             placeholder="Send a message"
-            onKeyDown={e => enterInput(e)}
+            onKeyDown={enterInput}
           />
         </div>
         <IconButton
