+++ src/modules/chats/components/sendMessage/index.types.ts
@@ -1,0 +1,5 @@
+export interface SendMessageProps {
+  recipientId?: string;
+  onSend?: (message: string) => void;
+  handleCreateChat?: (recipientId: string, text: string) => void;
+}
